Quarantine vertical grid to filename2 templates
The app store version doesn't support the vertical grid yet, but we want to go
live with the advertising changes. This makes it so the current app store
version is fully compatible with the templates.

The filename2 versions should be moved back to replace the filename versions
after users have been force upgraded to 7.7.